worker/requirements.txt
transformers
finnhub-python
firebase-admin
--extra-index-url https://download.pytorch.org/whl/cpu
torch
